Refactor structure of packages See #29
