+++ src/main/java/pl/pawel/line/model/Line.java
@@ -1,8 +1,8 @@
-package pl.pawel.model.line;
+package pl.pawel.line.model;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import pl.pawel.model.operatingControlPoint.OperatingControlPoint;
+import pl.pawel.operatingControlPoint.model.OperatingControlPoint;
 
 import javax.persistence.*;
 import java.util.Set;
