+++ src/main/java/pl/pawel/line/repository/LineRepository.java
@@ -1,9 +1,9 @@
-package pl.pawel.repository;
+package pl.pawel.line.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import pl.pawel.model.loadingPoint.LoadingPoint;
+import pl.pawel.line.model.Line;
 
 @RepositoryRestResource
-public interface LoadingPointRepository extends JpaRepository<LoadingPoint, Long> {
+public interface LineRepository extends JpaRepository<Line, Long> {
 }
