+++ src/main/java/pl/pawel/operatingControlPoint/repository/OperatingControlPointRepository.java
@@ -1,8 +1,8 @@
-package pl.pawel.repository;
+package pl.pawel.operatingControlPoint.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import pl.pawel.model.operatingControlPoint.OperatingControlPoint;
+import pl.pawel.operatingControlPoint.model.OperatingControlPoint;
 
 @RepositoryRestResource
 public interface OperatingControlPointRepository extends JpaRepository<OperatingControlPoint, Long> {
