+++ src/main/java/pl/pawel/railwayDepartment/repository/RailwayDepartmentRepository.java
@@ -1,8 +1,8 @@
-package pl.pawel.repository;
+package pl.pawel.railwayDepartment.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import pl.pawel.model.railwayDepartment.RailwayDepartment;
+import pl.pawel.railwayDepartment.model.RailwayDepartment;
 
 @RepositoryRestResource
 public interface RailwayDepartmentRepository extends JpaRepository<RailwayDepartment, Long> {
