--- src/main/java/pl/pawel/repository/EdgeRepository.java
@@ -1,9 +1,0 @@
-package pl.pawel.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import pl.pawel.model.edge.Edge;
-
-@RepositoryRestResource
-public interface EdgeRepository extends JpaRepository<Edge, Long> {
-}
